implement example flyweight pattern in python
This is a nifty one, the key for me was piecing together that, although we try to save memory by reusing object, we still have to create a new object for each new instance. It just that we're moving all state than CAN be shared (intrinsic) out of each instance and sharing that state. Theoretically, and niavely, we could translate the document now and all instance of 'a' could become '诶'